modelos, Dao, Services, Controllers

+++ src/main/java/com/apirest/efi/models/dao/CalendarioDao.java
@@ -3,6 +3,6 @@ package com.apirest.efi.models.dao;
 import com.apirest.efi.models.entity.Calendario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CalendarioDao extends JpaRepository<Calendario, Long> {
+public interface CalendarioDao extends JpaRepository<Calendario, Integer> {
 
 }
