+++ src/main/java/com/apirest/efi/models/dao/MunicipioDao.java
@@ -3,6 +3,6 @@ package com.apirest.efi.models.dao;
 import com.apirest.efi.models.entity.Municipio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MunicipioDao extends JpaRepository<Municipio, Long> {
+public interface MunicipioDao extends JpaRepository<Municipio, Integer> {
 
 }
